Restructure agent
The agent can now take an optional argument consisting of the path to a
config file. The config file is provided in JSON format and contains
rules for methods exclusion and inclusion:
 - `whitelist`: a list of strings specifying what classes and methods to
   instrument
 - `blacklist`: a list of strings specifying what classes and methods
   to never instrument
 - `excludeStandardJavaApi`: a boolean flag to exclude instrumentation
   on all the Java API standard classes. Defaults to `true`
 - `excludeStandardScalaApi`: a boolean flag to exclude instrumentation
   on all the Scala API standard classes. Defaults to `true`

The path to the config file can be specified given a full path or a path
that exists in the classpath. The agent will instrument methods that are
whitelisted and add the `Measured` annotation when it's not already
there.

Classes can also be annotated with `Measured` and all the
non-blacklisted methods within that class will be instrumented.

Also:
 - Move from Java 9 to Java 8
 - Rename module from `agent-instrumentation` to `agent`
 - Change base package from
   `tech.dario.dissertation.agent.instrumentation` to
   `tech.dario.timecomplexityanalysis.agent`
 - Rename `MeasuredClassTransformer` to `MeasuringClassFileTransformer`
 - Rename `MetricAgent` to `MeasuringAgent`