Change recording logic
**Problem**
When the agent instruments methods, it adds a call to
`Thread.currentThread().getStackTrace()` which takes 12-13 μs. As a
result the algorithm under test takes about twice the time when
instrumented compared to when it is not.

**Solution**
As there is no way to make `getStackTrace` faster, the instrumentation
logic had to be changed altogether. The new logic notifies the recorder
when a method is started and when it is finished, passing the method's
long name along. The recorder simply stores these method actions in a
list that grows as the algorithm under test runs. When the algorithm is
done running, the list is transformed into a call tree. The logic can
determine the stack trace based on the method actions that have been
received  so that there is no need for calling `getStackTrace` at any
time.

**Example**
The following method actions list:
```
MethodStarted("tech.dario.test.method1(String)", 0)
MethodStarted("tech.dario.test.method2(Integer)", 100)
MethodFinished("tech.dario.test.method2(Integer)", 300)
MethodFinished("tech.dario.test.method1(String)", 750)
```
shows `method1` calling `method2`, with `method1` taking 750 ns and
`method2` taking 200 ms.