Batch messages in time-recorder-akka
**Problem**
Sending a message to an Akka actor has some overhead. The message is put
inside an envelope and sits in the mailbox of the receiving actor. A lot
of envelopes can use a lot of memory, making the Garbage Collector
intervene stopping all thread for a moment.

**Solution**
Batch the messages in an array of a defined size X. This way there only
one envelope per X recordings.